Split summary.py to four files.
